feat: Switch to nalgebra and stl_io
I told the bot that TRUCK doesn't support Linux yet and at first it
chose cadquery_rs, but that doesn't exist. It then changed to using
this nalgebra and stl_io and this does compile and runs and creates
solenoid_spindle.stl, bravo. But it's not a spindle but still pretty
good.

There is a warning, the `use nalgebra as na;` statement is an unused_import
so I've commented that out for now.